Move Singleton definition to abstraction layer
